Add files via upload
The Photon Filter is a Mathematica .nb document, it uses the data from a binary file extracted from a open-source software which is Tonatiuh. Tonatiuh is used to make simulations of Thermosolar Systems. In this case, a Parabolic Dish System with a volumetric pressurized cylindrical receiver is modelled. The goal of this program is to know how many rays achieve the receiver in order to measure the power that can be expected. Once the power is known, the efficiency of the system can be measured.